fixes some type issues with ts
