+++ lib/add-project.js
@@ -1,6 +1,9 @@
 const { collectAndSaveProject } = require("./init/project");
 const projectsToText = require("./utils/projectsToText");
-const getSelectedProjects = require("./utils/getSelectedProjects");
+const {
+  getSelectedProjects,
+  getIsUsingComponents,
+} = require("./utils/getSelectedProjects");
 
 function quit(exitCode = 2) {
   console.log("Project selection was not updated.");
