+++ lib/remove-project.js
@@ -1,7 +1,7 @@
 const config = require("./config");
 const consts = require("./consts");
 const output = require("./output");
-const getSelectedProjects = require("./utils/getSelectedProjects");
+const { getSelectedProjects } = require("./utils/getSelectedProjects");
 const promptForProject = require("./utils/promptForProject");
 
 async function removeProject() {
